requirements.txt
opencv-contrib-python==4.8.0.76
numpy==1.26.4
Pillow==10.0.0 